feat: 添加面具

+++ eros-db/src/main/java/com/open/ai/eros/db/mysql/hr/entity/AmPosition.java
@@ -84,7 +84,7 @@ public class AmPosition implements Serializable {
     /**
      * ai助手id
      */
-    private Integer aiAssitantId;
+    private Long aiAssitantId;
 
     /**
      * 是否开放职位
