+++ hr-ai/src/main/java/com/open/hr/ai/bean/vo/AmPositionVo.java
@@ -84,7 +84,7 @@ public class AmPositionVo {
     /**
      * ai助手id
      */
-    private Integer aiAssitantId;
+    private Long aiAssitantId;
 
     /**
      * 是否开放职位
